+++ .github/workflows/main.yml
@@ -15,7 +15,8 @@ jobs:
         uses: actions/checkout@v2
       - name: Run script
         run: |
-          pip install black isort keyboard
+          pip install black isort pytest keyboard
           ls -la
           black --check .
           isort --profile black --check .
+          py.test tests
