+++ README.md
@@ -1,0 +1,31 @@
+**Симулятор робота перемещающегося по полю**
+
+Консольное интерактивное приложение, принимающее простой набор команд для управления роботом в рамках некоторого поля произвольного размера.
+Запуск приложения осуществляется с помощью файла initialization.py
+
+**Входные данные**
+
+Программа запускается из консоли и принимает поочередно одну из команд на вход
+1. Размеры поля
+2. Тип робота (на выбор предложено 3 формы)
+3. Количество препятствий
+4. Далее управление происходит с помощью клавиш (список команд с описанием высветится в консоли)
+5. esc - команда для завершения программы
+
+**Функциональность**
+1. Программа запускается из консоли и принимает поочередно одну из команд на вход
+2. Возможные команды: вверх/вниз, влево/вправо, поворот на 90/180
+3. Робот перемещается по полю X на Y, которые можно задать при запуске программы
+4. При запуске программы на поле генерируется N препятствий (количество указывается пользователем) 
+   разных форм(точка, \, квадрат), расположенных случайно: через них робот проехать не может
+5. Робот стартует по центру, после каждой команды докладывает о своей позиции: 
+   (X, Y), куда повернут;
+   откуда приехал
+6. Может быть выбор формы робота
+7. Отображение текущей позиции робота выдается в ascii графике с отрисовкой направления 
+   и близлежащих объектов в некотором радиусе
+8. Для более удобной работы с программой задействованы ascii цвета для вывода в терминал
+9. Хранение и экспорт результирующего маршрута робота в json-файл
+10. Для основных методов приложения написан набор тестов 
+
+
