+++ final_task/path.json
@@ -1,0 +1,1 @@
+{"1": [[5, 6], [5, 4], [5, 5, "↑"], [6, 5], [4, 5]], "2": [[5, 5], [5, 3], [5, 4, "↑"], [6, 4], [4, 4]]}
